Make Announcements More Consistent with Grid Below
This commit attempts to make the announcements component more consistent with the grid element below by utilizing grid instead of flexbox for spacing. It switches the padding of the one-box-grid and announcements.scss to be approximately the same so we get the same total width of the grid elements. We also retreat from our percent based aproach to use max-width 100% to limit to the calculated column sizes.