Javascript Refactor
This commit adds an .eslint configuration to the airbnb style guide to give eslint guidance on how to lint.

It also restructures our javascript to follow the entry point pattern enumerated at https://github.com/rails/webpacker/blob/master/docs/folder-structure.md.

Finally remove the videoContainer javascript that we do not plan to use in the weigh in redesign.