Fix Large Prompt Text Position
Here we modify the large prompt styling to share the styling rules of the stories so we have similar padding. We also add rules to center the large prompt text using flexbox to match the Adobe XD design. This also conveniently resolves an edge case for calculating the height of the entire outer stories container.